fix(HCScanner): orient='index'

+++ makiflow/augmentation/segmentation/balancing/utils.py
@@ -4,12 +4,12 @@ import numpy as np
 def hcv_to_num(bin_vec):
     num = 0
     for i in range(len(bin_vec)):
-        num += 2**i * bin_vec[i]
+        num += int(2**i * bin_vec[i])
     return num
 
 
 def to_hc_vec(num_classes, classes):
-    vec = np.zeros(num_classes)
+    vec = np.zeros(num_classes, dtype=np.uint8)
     vec[classes] = 1
     return vec
